chromium/third_party/blink/web_tests/http/tests/security/isolatedWorld/bypass-main-world-csp-font-src-expected.txt

ALERT: With lax isolated world CSP
ALERT: With strict isolated world CSP
This test ensures that scripts run in isolated worlds marked with their own Content Security Policy aren't affected by the page's font-src directive.
